Determine bottommost visible view for wrapped containers correctly.
The previous expression (offset + wrapperTop) was only correct when the wrapped
container was positioned at the top of the document. If the container was
further down from the top of the document than the height of the last row,
the last row would always be cloaked. The further down the container started,
the more of the bottommost rows would always be cloaked.

This was because "viewTop" was computed using document-relative coordinates,
but was compared to a value (viewportBottom) that was clamped to the height
of the container's content.

This change uses a coordinate for viewTop which is relative to the scrolling
container.