.position() is already relative to the scroll container.
Incorporating wrapperTop into the calculated viewBottom shifts the computed
bounds for all the views such that the first row is always found to be the
topmost visible. Top put it another way, it prevents any of the views that
scroll above the top from ever being cloaked when you're using a wrapper
element.